Getting all inputs for AES GCM from the FPGA I/O pins.
Due to this there is no more constant propogation of IV
and the AES cipher for counter blocks is not getting
precomputed. Because of the maximum input size that the
implementation can support is 128 bit plain text and AAD
data.

TODO: This can be improved by loading larger inputs into
a register and then accessing plain text and AAD blockk by block
in different cycles. This idea needs to be tested.